x/story/keeper/msg_server_event_test.go
package keeper_test

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"

	"github.com/web-tree/story/x/story/types"
)

func TestEventMsgServerCreate(t *testing.T) {
	srv, ctx := setupMsgServer(t)
	creator := "A"
	for i := 0; i < 5; i++ {
		resp, err := srv.CreateEvent(ctx, &types.MsgCreateEvent{Creator: creator})
		require.NoError(t, err)
		require.Equal(t, i, int(resp.Id))
	}
}

func TestEventMsgServerUpdate(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgUpdateEvent
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdateEvent{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateEvent{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateEvent{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)
			_, err := srv.CreateEvent(ctx, &types.MsgCreateEvent{Creator: creator})
			require.NoError(t, err)

			_, err = srv.UpdateEvent(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}

func TestEventMsgServerDelete(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgDeleteEvent
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeleteEvent{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeleteEvent{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "KeyNotFound",
			request: &types.MsgDeleteEvent{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)

			_, err := srv.CreateEvent(ctx, &types.MsgCreateEvent{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeleteEvent(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
